+++ app/components/EntryEditor/SearchResults.tsx
@@ -5,12 +5,12 @@ import {
 } from "~/routes/search/searchResultEnum"
 import { type AllSearchResults } from "~/models/search.server"
 import Button from "../elements/LinksAndButtons/Button"
-import SearchResultCanadianism from "../search/SearchResultCanadianism"
-import SearchResultEntries from "../search/SearchResultEntries"
-import SearchResultFistNotes from "../search/SearchResultFistNotes"
-import SearchResultMeanings from "../search/SearchResultMeanings"
-import SearchResultQuotations from "../search/SearchResultQuotations"
-import SearchResultUsageNotes from "../search/SearchResultUsageNotes"
+import SearchResultCanadianism from "../search/Results/SearchResultCanadianism"
+import SearchResultEntries from "../search/Results/SearchResultEntries"
+import SearchResultFistNotes from "../search/Results/SearchResultFistNotes"
+import SearchResultMeanings from "../search/Results/SearchResultMeanings"
+import SearchResultQuotations from "../search/Results/SearchResultQuotations"
+import SearchResultUsageNotes from "../search/Results/SearchResultUsageNotes"
 import type JSXNode from "~/types/JSXNode"
 import PaginationControl from "../bank/PaginationControl"
 import { DEFAULT_PAGE_SIZE } from "~/utils/pageSize"
