+++ app/components/search/Results/SearchResultQuotations.tsx
@@ -1,6 +1,6 @@
 import type { Quotation } from "~/models/search/getSearchResultQuotations"
-import SanitizedTextSpan from "../Entry/Common/SanitizedTextSpan"
-import { Link } from "../elements/LinksAndButtons/Link"
+import SanitizedTextSpan from "../../Entry/Common/SanitizedTextSpan"
+import { Link } from "../../elements/LinksAndButtons/Link"
 
 interface SearchResultProps {
   text: string
